REF: refactor and fix structured data
