+++ src/app/about/page.tsx
@@ -1,6 +1,5 @@
 import { Files, LinkedinLogo } from '@phosphor-icons/react/dist/ssr'
 import clsx from 'clsx'
-import { WebPage, WithContext } from 'schema-dts'
 
 import { Card } from '@/components/Card'
 import { CardGrid } from '@/components/CardGrid'
@@ -10,15 +9,12 @@ import { PageSection } from '@/components/PageSection'
 import { StructuredDataScript } from '@/components/StructuredDataScript'
 
 import { createMetadata } from '@/utils/createMetadata'
-import { generateWebPageStructuredData } from '@/utils/structuredData'
+import { generateAboutPageStructuredData } from '@/utils/structuredData'
 
 import { attributes } from '@/content/pages/about.md'
 
 import { PATHS } from '@/constants/paths'
-import {
-  ORGANIZATION_NAME,
-  FILECOIN_FOUNDATION_URLS,
-} from '@/constants/siteMetadata'
+import { FILECOIN_FOUNDATION_URLS } from '@/constants/siteMetadata'
 
 import { advisorsData } from './data/advisorsData'
 import { boardMembersData } from './data/boardMembersData'
@@ -29,39 +25,12 @@ const { header, seo } = attributes
 
 export const metadata = createMetadata({ seo, path: PATHS.ABOUT.path })
 
-const aboutPageBaseData = generateWebPageStructuredData({
-  title: seo.title,
-  description: seo.description,
-  path: PATHS.ABOUT.path,
-})
-
-const aboutPageStructuredData: WithContext<WebPage> = {
-  ...aboutPageBaseData,
-  about: {
-    '@type': 'Organization',
-    name: ORGANIZATION_NAME,
-    contactPoint: [
-      {
-        '@type': 'ContactPoint',
-        contactType: 'Media and collaboration inquiries',
-        email: FILECOIN_FOUNDATION_URLS.email,
-      },
-      {
-        '@type': 'ContactPoint',
-        contactType: 'Ecosystem grants inquiries',
-        email: FILECOIN_FOUNDATION_URLS.grants.email,
-      },
-    ],
-  },
-  sameAs: Object.values(FILECOIN_FOUNDATION_URLS.social).map(
-    (link) => link.href,
-  ),
-}
-
 export default function About() {
   return (
     <PageLayout>
-      <StructuredDataScript structuredData={aboutPageStructuredData} />
+      <StructuredDataScript
+        structuredData={generateAboutPageStructuredData(seo)}
+      />
       <PageHeader
         title={header.title}
         description={header.description}
