+++ src/app/events/page.tsx
@@ -1,7 +1,6 @@
 import dynamic from 'next/dynamic'
 
 import { MagnifyingGlass } from '@phosphor-icons/react/dist/ssr'
-import { WebPage, WithContext } from 'schema-dts'
 
 import { useCategory } from '@/hooks/useCategory'
 import { usePagination } from '@/hooks/usePagination'
@@ -28,12 +27,11 @@ import { getCategorySettings } from '@/utils/categoryUtils'
 import { createMetadata } from '@/utils/createMetadata'
 import { formatDate } from '@/utils/formatDate'
 import { getEventsData } from '@/utils/getEventData'
-import { generateWebPageStructuredData } from '@/utils/structuredData'
+import { generateEventsPageStructuredData } from '@/utils/structuredData'
 
 import { attributes } from '@/content/pages/events.md'
 
 import { PATHS } from '@/constants/paths'
-import { BASE_URL } from '@/constants/siteMetadata'
 
 import { getInvolvedData } from './data/getInvolvedData'
 
@@ -52,31 +50,6 @@ const { featured_entry: featuredEventSlug, seo } = attributes
 const featuredEvent = events.find((event) => event.slug === featuredEventSlug)
 
 export const metadata = createMetadata({ seo, path: PATHS.EVENTS.path })
-
-const eventsPageBaseData = generateWebPageStructuredData({
-  title: seo.title,
-  description: seo.description,
-  path: PATHS.EVENTS.path,
-})
-
-const eventsPageStructuredData: WithContext<WebPage> = {
-  ...eventsPageBaseData,
-  mainEntity: {
-    '@type': 'ItemList',
-    itemListElement: events.slice(0, 5).map((event, index) => ({
-      '@type': 'ListItem',
-      position: index + 1,
-      item: {
-        '@type': 'Event',
-        name: event.title,
-        startDate: event.startDate,
-        endDate: event.endDate,
-        description: event.description,
-        url: `${event.externalLink?.href} || ${BASE_URL}${PATHS.EVENTS.path}/${event.slug}`,
-      },
-    })),
-  },
-}
 
 function prepareMetaData(
   startDate: EventData['startDate'],
@@ -145,7 +118,9 @@ export default function Events({ searchParams }: Props) {
 
   return (
     <PageLayout>
-      <StructuredDataScript structuredData={eventsPageStructuredData} />
+      <StructuredDataScript
+        structuredData={generateEventsPageStructuredData(events, seo)}
+      />
       <PageHeader
         isFeatured
         title={featuredEvent.title}
