+++ src/app/privacy-policy/page.tsx
@@ -2,7 +2,7 @@ import { PageHeader } from '@/components/PageHeader'
 import { StructuredDataScript } from '@/components/StructuredDataScript'
 
 import { createMetadata } from '@/utils/createMetadata'
-import { generateWebPageStructuredData } from '@/utils/structuredData'
+import { generatePrivacyPolicyStructuredData } from '@/utils/structuredData'
 
 import { attributes, react as Content } from '@/content/pages/privacy-policy.md'
 
@@ -12,18 +12,14 @@ const { header, seo } = attributes
 
 export const metadata = createMetadata({ seo, path: PATHS.PRIVACY_POLICY.path })
 
-const policyPageBaseData = generateWebPageStructuredData({
-  title: seo.title,
-  description: seo.description,
-  path: PATHS.PRIVACY_POLICY.path,
-})
-
 export default function PrivacyPolicy() {
   const { title } = header
 
   return (
     <article>
-      <StructuredDataScript structuredData={policyPageBaseData} />
+      <StructuredDataScript
+        structuredData={generatePrivacyPolicyStructuredData(seo)}
+      />
       <PageHeader.Title>{title}</PageHeader.Title>
       <section className="prose mt-6">
         <Content />
